+++ rocketpy/rocket/rocket.py
@@ -885,7 +885,7 @@ class Rocket:
 
         See Also
         --------
-        :ref:`positions`
+        :ref:`addsurface`
 
         Returns
         -------
@@ -951,7 +951,7 @@ class Rocket:
 
         See Also
         --------
-        :ref:`positions`
+        :ref:`addsurface`
 
         Returns
         -------
@@ -1015,10 +1015,16 @@ class Rocket:
             By tail position, understand the point belonging to the tail which
             is highest in the rocket coordinate system (i.e. the point
             closest to the nose cone).
+        radius : int, float, optional
+            Reference radius of the tail. This is used to calculate lift
+            coefficient. If None, which is default, the rocket radius will
+            be used.
+        name : string
+            Tail name. Default is "Tail".
 
         See Also
         --------
-        :ref:`positions`
+        :ref:`addsurface`
 
         Returns
         -------
@@ -1047,7 +1053,6 @@ class Rocket:
         along the rocket and its derivative of the coefficient of lift
         in respect to angle of attack.
 
-
         Parameters
         ----------
         length : int, float
@@ -1073,7 +1078,7 @@ class Rocket:
 
         See Also
         --------
-        :ref:`positions`
+        :ref:`addsurface`
 
         Returns
         -------
@@ -1259,6 +1264,10 @@ class Rocket:
             return the lift coefficient at that angle of attack.
             The tuple's second item is the unit of the angle of attack,
             accepting either "radians" or "degrees".
+
+        See Also
+        --------
+        :ref:`addsurface`
 
         Returns
         -------
@@ -1486,7 +1495,11 @@ class Rocket:
             return air_brakes
 
     def set_rail_buttons(
-        self, upper_button_position, lower_button_position, angular_position=45
+        self,
+        upper_button_position,
+        lower_button_position,
+        angular_position=45,
+        radius=None,
     ):
         """Adds rail buttons to the rocket, allowing for the calculation of
         forces exerted by them when the rocket is sliding in the launch rail.
@@ -1511,10 +1524,12 @@ class Rocket:
             relative to one of the other principal axis.
             Default value is 45 degrees, generally used in rockets with
             4 fins.
+        radius : int, float, optional
+            Fuselage radius where the rail buttons are located.
 
         See Also
         --------
-        :ref:`positions`
+        :ref:`addsurface`
 
         Returns
         -------
@@ -1525,7 +1540,8 @@ class Rocket:
         rail_buttons = RailButtons(
             buttons_distance=buttons_distance, angular_position=angular_position
         )
-        rail_buttons.rocket_radius = rail_buttons.rocket_radius or self.radius
+        self.rail_buttons = Components()
+        rail_buttons.rocket_radius = radius or self.radius
         self.rail_buttons.add(rail_buttons, lower_button_position)
         return rail_buttons